requirements.txt
streamlit==1.41.1
azure-cognitiveservices-speech==1.16.0
langchain==0.3.0
langchain-core~=0.3.15
langchain-openai~=0.2.6